Refactor agent package to use Repository Pattern for framework independence
This change eliminates direct dependencies on @liam-hq/db and Next.js specific features,
making the agent package usable by external tools like benchmarkers.

Key changes:
- Introduce Repository Pattern with SchemaRepository interface
- Create SupabaseSchemaRepository implementation
- Add factory functions for repository creation
- Update WorkflowState to include repositories dependency
- Remove direct database and Next.js dependencies
- Create helper functions in app package for Supabase integration
- Maintain backward compatibility for existing API
