+++ frontend/apps/app/app/(app)/app/design_sessions/[id]/page.tsx
@@ -1,6 +1,9 @@
 import type { PageProps } from '@/app/types'
 import { SessionDetailPage } from '@/components/SessionDetailPage'
-import { fetchDesignSessionData, fetchSchemaData } from '@liam-hq/agent'
+import {
+  fetchDesignSessionData,
+  fetchSchemaData,
+} from '@/utils/agentSupabaseHelper'
 import * as v from 'valibot'
 
 const paramsSchema = v.object({
