+++ frontend/apps/app/app/api/buildingSchemas/versions/route.ts
@@ -1,4 +1,5 @@
-import { createNewVersion, operationsSchema } from '@liam-hq/agent'
+import { createNewVersion } from '@/utils/agentSupabaseHelper'
+import { operationsSchema } from '@liam-hq/agent'
 import { type NextRequest, NextResponse } from 'next/server'
 import * as v from 'valibot'
