+++ frontend/internal-packages/agent/package.json
@@ -12,7 +12,6 @@
     "@liam-hq/db-structure": "workspace:*",
     "fast-json-patch": "3.1.1",
     "langfuse-langchain": "3.37.2",
-    "next": "15.3.2",
     "ts-pattern": "5.7.1",
     "valibot": "1.1.0"
   },
