+++ frontend/internal-packages/agent/src/index.ts
@@ -1,5 +1,3 @@
+export * from './repositories'
 export * from './chat/chatProcessor'
-export * from './utils/applyPatchOperations'
 export * from './utils/operationsSchema'
-export * from './utils/createNewVersion'
-export * from './utils/fetchSchemaData'
