+++ frontend/internal-packages/agent/src/repositories/index.ts
@@ -1,0 +1,12 @@
+export type {
+  SchemaRepository,
+  Repositories,
+  SchemaData,
+  DesignSessionData,
+  CreateVersionParams,
+  VersionResult,
+} from './types'
+
+export { SupabaseSchemaRepository } from './supabase'
+
+export { createSupabaseRepositories } from './factory'
