+++ frontend/internal-packages/agent/src/repositories/supabase.ts
@@ -1,0 +1,309 @@
+import type { SupabaseClientType } from '@liam-hq/db'
+import { schemaSchema } from '@liam-hq/db-structure'
+import { compare } from 'fast-json-patch'
+import * as v from 'valibot'
+import { applyPatchOperations } from '../utils/applyPatchOperations'
+import { operationsSchema } from '../utils/operationsSchema'
+import type {
+  CreateVersionParams,
+  DesignSessionData,
+  SchemaData,
+  SchemaRepository,
+  VersionResult,
+} from './types'
+
+const updateBuildingSchemaResultSchema = v.union([
+  v.object({
+    success: v.literal(true),
+    versionNumber: v.number(),
+  }),
+  v.object({
+    success: v.literal(false),
+    error: v.nullable(v.string()),
+  }),
+])
+
+/**
+ * Supabase implementation of SchemaRepository
+ */
+export class SupabaseSchemaRepository implements SchemaRepository {
+  private client: SupabaseClientType
+
+  constructor(client: SupabaseClientType) {
+    this.client = client
+  }
+
+  async getDesignSession(
+    designSessionId: string,
+  ): Promise<DesignSessionData | null> {
+    // Fetch design session with messages
+    const { data, error } = await this.client
+      .from('design_sessions')
+      .select(`
+        organization_id,
+        messages (
+          id,
+          content,
+          role,
+          user_id,
+          created_at,
+          updated_at,
+          organization_id,
+          design_session_id
+        )
+      `)
+      .eq('id', designSessionId)
+      .order('created_at', { ascending: true, referencedTable: 'messages' })
+      .single()
+
+    if (error || !data) {
+      console.error(
+        `Could not fetch design session data for ${designSessionId}:`,
+        error,
+      )
+      return null
+    }
+
+    return {
+      organization_id: data.organization_id,
+      messages: data.messages || [],
+    }
+  }
+
+  async getSchema(designSessionId: string): Promise<{
+    data: SchemaData | null
+    error: { message: string } | null
+  }> {
+    const buildingSchemaResult = await this.getBuildingSchema(designSessionId)
+    if (buildingSchemaResult.error || !buildingSchemaResult.data) {
+      return buildingSchemaResult
+    }
+
+    const { buildingSchema } = buildingSchemaResult.data
+
+    const versionsResult = await this.getSchemaVersions(buildingSchema.id)
+    if (versionsResult.error) {
+      return versionsResult
+    }
+
+    const { versions } = versionsResult.data
+    const currentSchema = this.buildCurrentSchema(buildingSchema, versions)
+    const latestVersionNumber = this.getLatestVersionNumber(versions)
+
+    return {
+      data: {
+        id: buildingSchema.id,
+        schema: currentSchema,
+        latestVersionNumber,
+      },
+      error: null,
+    }
+  }
+
+  private async getBuildingSchema(designSessionId: string) {
+    const { data: buildingSchema, error: buildingSchemaError } =
+      await this.client
+        .from('building_schemas')
+        .select('id, initial_schema_snapshot')
+        .eq('design_session_id', designSessionId)
+        .single()
+
+    if (buildingSchemaError || !buildingSchema) {
+      return {
+        data: null,
+        error: buildingSchemaError
+          ? { message: buildingSchemaError.message }
+          : null,
+      }
+    }
+
+    return { data: { buildingSchema }, error: null }
+  }
+
+  private async getSchemaVersions(buildingSchemaId: string) {
+    const { data: versions, error: versionsError } = await this.client
+      .from('building_schema_versions')
+      .select('number, patch')
+      .eq('building_schema_id', buildingSchemaId)
+      .order('number', { ascending: true })
+
+    if (versionsError) {
+      return {
+        data: null,
+        error: { message: versionsError.message },
+      }
+    }
+
+    return { data: { versions: versions || [] }, error: null }
+  }
+
+  private buildCurrentSchema(
+    buildingSchema: { initial_schema_snapshot: unknown },
+    versions: Array<{ number: number; patch: unknown }>,
+  ): Record<string, unknown> {
+    const currentSchema: Record<string, unknown> =
+      typeof buildingSchema.initial_schema_snapshot === 'object' &&
+      buildingSchema.initial_schema_snapshot !== null
+        ? JSON.parse(JSON.stringify(buildingSchema.initial_schema_snapshot))
+        : { tables: {}, relationships: {}, tableGroups: {} }
+
+    for (const version of versions) {
+      const patchParsed = v.safeParse(operationsSchema, version.patch)
+      if (patchParsed.success) {
+        applyPatchOperations(currentSchema, patchParsed.output)
+      } else {
+        console.warn(
+          `Invalid patch operations in version ${version.number}:`,
+          version.patch,
+        )
+      }
+    }
+
+    return currentSchema
+  }
+
+  private getLatestVersionNumber(versions: Array<{ number: number }>): number {
+    return versions.length > 0 ? Math.max(...versions.map((v) => v.number)) : 0
+  }
+
+  async createVersion(params: CreateVersionParams): Promise<VersionResult> {
+    const { buildingSchemaId, latestVersionNumber, patch } = params
+
+    const { data: buildingSchema, error } = await this.client
+      .from('building_schemas')
+      .select(`
+        id, organization_id, initial_schema_snapshot
+      `)
+      .eq('id', buildingSchemaId)
+      .maybeSingle()
+
+    if (!buildingSchema || error) {
+      throw new Error(`Failed to fetch building schema: ${error?.message}`)
+    }
+
+    // Get all previous versions to reconstruct the content
+    const { data: previousVersions, error: previousVersionsError } =
+      await this.client
+        .from('building_schema_versions')
+        .select('number, patch')
+        .eq('building_schema_id', buildingSchemaId)
+        .lte('number', latestVersionNumber)
+        .order('number', { ascending: true })
+
+    if (previousVersionsError) {
+      throw new Error(
+        `Failed to fetch previous versions: ${previousVersionsError.message}`,
+      )
+    }
+
+    const patchArrayHistory = previousVersions
+      ?.map((version) => {
+        const parsed = v.safeParse(operationsSchema, version.patch)
+        if (parsed.success) {
+          return parsed.output
+        }
+        return null
+      })
+      .filter((version) => version !== null)
+
+    // Reconstruct the base content (first version) from the initial schema snapshot
+    const baseContent: Record<string, unknown> =
+      typeof buildingSchema.initial_schema_snapshot === 'object'
+        ? JSON.parse(JSON.stringify(buildingSchema.initial_schema_snapshot))
+        : {}
+
+    // Apply all patches in order to get the current content
+    const currentContent: Record<string, unknown> = { ...baseContent }
+
+    // Apply all patches in order
+    for (const patchArray of patchArrayHistory) {
+      // Apply each operation in the patch to currentContent
+      applyPatchOperations(currentContent, patchArray)
+    }
+
+    // Now apply the new patch to get the new content
+    const newContent = JSON.parse(JSON.stringify(currentContent))
+    applyPatchOperations(newContent, patch)
+
+    // Validate the new schema structure before proceeding
+    const validationResult = v.safeParse(schemaSchema, newContent)
+    if (!validationResult.success) {
+      const errorMessages = validationResult.issues
+        .map((issue) => `${issue.path?.join('.')} ${issue.message}`)
+        .join(', ')
+      return {
+        success: false,
+        error: `Invalid schema after applying changes: ${errorMessages}`,
+      }
+    }
+
+    // Calculate reverse patch from new content to current content
+    const reversePatch = compare(newContent, currentContent)
+
+    // Get the latest version number for this schema
+    const { data: latestVersion, error: latestVersionError } = await this.client
+      .from('building_schema_versions')
+      .select('number')
+      .eq('building_schema_id', buildingSchemaId)
+      .order('number', { ascending: false })
+      .limit(1)
+      .maybeSingle()
+
+    if (latestVersionError) {
+      throw new Error(
+        `Failed to get latest version: ${latestVersionError.message}`,
+      )
+    }
+
+    // Get the actual latest version number
+    const actualLatestVersionNumber = latestVersion ? latestVersion.number : 0
+
+    // Check if the expected version number matches the actual latest version number
+    if (latestVersionNumber !== actualLatestVersionNumber) {
+      // Version conflict detected
+      return {
+        success: false,
+        error:
+          'Version conflict: The schema has been modified since you last loaded it',
+      }
+    }
+
+    const { data, error: rpcError } = await this.client.rpc(
+      'update_building_schema',
+      {
+        p_schema_id: buildingSchemaId,
+        p_schema_schema: newContent,
+        p_schema_version_patch: JSON.parse(JSON.stringify(patch)),
+        p_schema_version_reverse_patch: JSON.parse(
+          JSON.stringify(reversePatch),
+        ),
+        p_latest_schema_version_number: actualLatestVersionNumber,
+      },
+    )
+
+    const parsedResult = v.safeParse(updateBuildingSchemaResultSchema, data)
+
+    if (rpcError) {
+      return {
+        success: false,
+        error: rpcError.message,
+      }
+    }
+
+    if (parsedResult.success) {
+      if (parsedResult.output.success) {
+        return {
+          success: true,
+        }
+      }
+      return {
+        success: false,
+        error: parsedResult.output.error,
+      }
+    }
+    return {
+      success: false,
+      error: 'Invalid response from server',
+    }
+  }
+}
